making video player & auth

+++ controllers/videoControllers.js
@@ -42,53 +42,66 @@ export const postUpload = async (req, res) => {
     const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
+        description,
+        creator: req.user.id
     });
-    console.log(newVideo);
+    req.user.videos.push(newVideo._id);
+    req.user.save();
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
 export const videoDetail = async (req, res) => {
     const { params : {id}} = req;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate('creator');
     try {     
         res.render("videoDetail", { pageTitle: video.title , video });
     } catch (error) {
         res.redirect(routes.home);
     }
 }
+// Edit Video
 
-export const editVideo = async (req, res) => {
-    if (req.method === "GET") {
-        const {
-            files: { id }
-          } = req;
-          try {
-            const video = await Video.findById(id);
-            res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
-          } catch (error) {
-            res.redirect(routes.home);
-          } 
-    } else if (req.method === "POST")
-    {
-        const {
-            params: { id },
-            body: { title, description }
-          } = req;
-          try {
-            await Video.findOneAndUpdate({ _id: id }, { title, description });
-            res.redirect(routes.videoDetail(id));
-          } catch (error) {
-            res.redirect(routes.home);
-          }        
+export const getEditVideo = async (req, res) => {
+    const {
+      params: { id }
+    } = req;
+    try {
+      const video = await Video.findById(id);
+      if (String(video.creator) !== req.user.id) {
+        throw Error();
+      } else {
+        res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+      }
+    } catch (error) {
+        console.log(error);
+      res.redirect(routes.home);
     }
-}
+  };
+  
+  export const postEditVideo = async (req, res) => {
+    const {
+      params: { id },
+      body: { title, description }
+    } = req;
+    try {
+      await Video.findOneAndUpdate({ _id: id }, { title, description });
+      res.redirect(routes.videoDetail(id));
+    } catch (error) {
+      res.redirect(routes.home);
+    }
+  };
+  
 export const deleteVideo = async (req, res) => {
     const {
         params: { id }
     } =req;
     try {
+    const video = await Video.findById(id);
+      if (video.creator !== req.user.id) {
+        throw Error();
+      } else {
         await Video.findOneAndRemove({_id:id});
+      }
     } catch (error) {
         console.log(error);
     }
